maint: Split cli.py into separate sub modules for separation of concerns. Also update click-aliases for add_command-method support.
